Moved TableAnnotationRequest to dataset.py
